+++ integration_tests/suite/test_voicemail_transcription.py
@@ -3,7 +3,16 @@
 
 import uuid
 
-from hamcrest import assert_that, has_entries, has_item, has_items, is_, none
+from hamcrest import (
+    assert_that,
+    contains_exactly,
+    has_entries,
+    has_item,
+    has_items,
+    is_,
+    none,
+    only_contains,
+)
 from wazo_test_helpers import until
 
 from .helpers.base import IntegrationTest
@@ -203,7 +212,6 @@ class TestVoicemailTranscriptionEnrichment(RealAsteriskIntegrationTest):
         )
         self.confd.set_user_voicemails({user_uuid: [voicemail]})
         self.confd.set_voicemails(voicemail)
-
         # No transcriptions set in call-logd
         calld = self.make_user_calld(user_uuid, tenant_uuid=VALID_TENANT)
         result = calld.voicemails.list_voicemail_messages_from_user(
@@ -490,6 +498,169 @@ class TestVoicemailTranscriptionEnrichment(RealAsteriskIntegrationTest):
                         ),
                     ),
                 ),
+            ),
+        )
+
+    def test_list_admin_messages_with_transcription(self):
+        user_uuid = str(uuid.uuid4())
+        voicemail_id = 111
+        message_id = '1724107750-00000001'  # Present in Docker volume
+        voicemail = MockVoicemail(
+            voicemail_id,
+            '8000',
+            'admin-voicemail',
+            'default',
+            tenant_uuid=VALID_TENANT,
+        )
+        self.confd.set_voicemails(voicemail)
+
+        self.call_logd.set_transcriptions(
+            [
+                {
+                    'message_id': message_id,
+                    'voicemail_id': voicemail_id,
+                    'tenant_uuid': VALID_TENANT,
+                    'transcription_text': 'Admin list transcription',
+                    'provider_id': 'openai/whisper-1',
+                    'language': 'en',
+                    'duration': 19.0,
+                },
+            ]
+        )
+
+        calld = self.make_user_calld(user_uuid, tenant_uuid=VALID_TENANT)
+        result = calld.voicemails.list_voicemail_messages()
+
+        items = result['items']
+        msg = next(m for m in items if m['id'] == message_id)
+        assert_that(
+            msg.keys(),
+            only_contains(
+                'id',
+                'caller_id_name',
+                'caller_id_num',
+                'duration',
+                'timestamp',
+                'empty',
+                'transcribed',
+                'voicemail',
+                'folder',
+            ),
+        )
+        assert_that(msg['transcribed'], is_(True))
+
+    def test_list_admin_messages_without_transcription(self):
+        user_uuid = str(uuid.uuid4())
+        voicemail_id = 111
+        message_id = '1724107750-00000001'  # Present in Docker volume
+        voicemail = MockVoicemail(
+            voicemail_id,
+            '8000',
+            'admin-voicemail',
+            'default',
+            tenant_uuid=VALID_TENANT,
+        )
+        self.confd.set_voicemails(voicemail)
+
+        calld = self.make_user_calld(user_uuid, tenant_uuid=VALID_TENANT)
+        result = calld.voicemails.list_voicemail_messages()
+
+        items = result['items']
+        msg = next(m for m in items if m['id'] == message_id)
+        assert_that(
+            msg.keys(),
+            only_contains(
+                'id',
+                'caller_id_name',
+                'caller_id_num',
+                'duration',
+                'timestamp',
+                'empty',
+                'transcribed',
+                'voicemail',
+                'folder',
+            ),
+        )
+        assert_that(msg['transcribed'], is_(False))
+
+    def test_list_admin_messages_filter_transcribed(self):
+        user_uuid = str(uuid.uuid4())
+        voicemail_id_1 = 111
+        voicemail_id_2 = 222
+        transcribed_msg = '1724107750-00000001'  # default/8000
+        non_transcribed_msg = '1724436688-00000001'  # default/8001
+        vm1 = MockVoicemail(
+            voicemail_id_1,
+            '8000',
+            'vm-transcribed',
+            'default',
+            tenant_uuid=VALID_TENANT,
+        )
+        vm2 = MockVoicemail(
+            voicemail_id_2,
+            '8001',
+            'vm-not-transcribed',
+            'default',
+            tenant_uuid=VALID_TENANT,
+        )
+        self.confd.set_voicemails(vm1, vm2)
+
+        self.call_logd.set_transcriptions(
+            [
+                {
+                    'message_id': transcribed_msg,
+                    'voicemail_id': voicemail_id_1,
+                    'tenant_uuid': VALID_TENANT,
+                    'transcription_text': 'Hello',
+                    'provider_id': 'openai/whisper-1',
+                    'language': 'en',
+                    'duration': 19.0,
+                },
+            ]
+        )
+
+        calld = self.make_user_calld(user_uuid, tenant_uuid=VALID_TENANT)
+        # default/8001 has 3 more messages, total = 4
+
+        # no filter: both transcribed and non-transcribed included
+        result = calld.voicemails.list_voicemail_messages()
+        assert_that(
+            result,
+            has_entries(
+                items=has_items(
+                    has_entries(id=transcribed_msg, transcribed=True),
+                    has_entries(id=non_transcribed_msg, transcribed=False),
+                ),
+                total=4,
+                filtered=4,
+            ),
+        )
+
+        # transcribed=True: only transcribed messages
+        result = calld.voicemails.list_voicemail_messages(transcribed=True)
+        assert_that(
+            result,
+            has_entries(
+                items=contains_exactly(
+                    has_entries(id=transcribed_msg, transcribed=True),
+                ),
+                total=4,
+                filtered=1,
+            ),
+        )
+
+        # transcribed=False: only non-transcribed messages
+        result = calld.voicemails.list_voicemail_messages(transcribed=False)
+        assert_that(
+            result,
+            has_entries(
+                items=contains_exactly(
+                    has_entries(id='1724436688-00000001', transcribed=False),
+                    has_entries(id='1724436755-00000002', transcribed=False),
+                    has_entries(id='1724436995-00000003', transcribed=False),
+                ),
+                total=4,
+                filtered=3,
             ),
         )