+++ wazo_calld/plugins/voicemails/http.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2025 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2026 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import re
@@ -21,6 +21,8 @@ from .exceptions import (
 )
 from .schemas import (
     VALID_GREETINGS,
+    voicemail_admin_messages_get_schema,
+    voicemail_admin_messages_schema,
     voicemail_folder_schema,
     voicemail_greeting_copy_schema,
     voicemail_message_schema,
@@ -358,6 +360,20 @@ def _validate_greeting(greeting):
     if greeting in VALID_GREETINGS:
         return greeting
     raise NoSuchVoicemailGreeting(greeting)
+
+
+class VoicemailMessagesResource(AuthResource):
+    def __init__(self, voicemails_service):
+        self._voicemails_service = voicemails_service
+
+    @required_acl('calld.voicemails.messages.read')
+    def get(self):
+        params = voicemail_admin_messages_get_schema.load(request.args.to_dict())
+        tenant = Tenant.autodetect()
+
+        result = self._voicemails_service.get_tenant_messages(tenant.uuid, **params)
+
+        return voicemail_admin_messages_schema.dump(result)
 
 
 class UserVoicemailMessagesResource(AuthResource):